Move safepath
