+++ apps/provisioning/pkg/safepath/dir_test.go
