+++ apps/provisioning/pkg/safepath/walk.go
