+++ pkg/registry/apis/provisioning/repository/git/staged.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/grafana/grafana/apps/provisioning/pkg/safepath"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
-	"github.com/grafana/grafana/pkg/registry/apis/provisioning/safepath"
 	"github.com/grafana/nanogit"
 )
