+++ pkg/registry/apis/provisioning/resources/tree.go
@@ -8,8 +8,8 @@ import (
 
 	folders "github.com/grafana/grafana/apps/folder/pkg/apis/folder/v1beta1"
 	provisioning "github.com/grafana/grafana/apps/provisioning/pkg/apis/provisioning/v0alpha1"
+	"github.com/grafana/grafana/apps/provisioning/pkg/safepath"
 	"github.com/grafana/grafana/pkg/apimachinery/utils"
-	"github.com/grafana/grafana/pkg/registry/apis/provisioning/safepath"
 )
 
 // FolderTree contains the entire set of folders (at a given snapshot in time) of the Grafana instance.
